requirements.txt
pandas>=2.0.0
numpy>=1.24.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
shap>=0.42.0
streamlit>=1.28.0
plotly>=5.17.0
faker>=19.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
joblib>=1.3.0
seaborn>=0.12.0

